requirements.txt
pandas>=2.2
openpyxl>=3.1
pyyaml>=6.0
pypdf>=5.0
tabulate>=0.9
